Refactor `Patch` into a separate module
Including `DownloadInfoTooltip`, and `Upload`.
