+++ WebRandomizer/ClientApp/src/patch/DownloadInfoTooltip.jsx
