+++ WebRandomizer/ClientApp/src/patch/Patch.jsx
